fix(core): move ai-categorizer cache out of finance (PRD-097, #2518)
The Claude-Haiku disk cache for AI calls is cross-cutting AI infra,
not a finance concern. Moving it from
`apps/pops-api/src/modules/finance/imports/lib/ai-categorizer-cache.ts`
to `apps/pops-api/src/modules/core/ai-usage/cache.ts`:

- core/ai-usage/router.ts now imports from a local sibling instead
  of crossing into the finance module (drops the
  `no-cross-api-module-import` violation baselined in PRD-097)
- finance/imports/lib/ai-categorizer.ts + ai-categorizer-api.ts
  import the cache from `core/ai-usage` (allowed: domain modules
  may depend on core)

Baseline shrinks from 14 to 13 known violations. All 74
ai-categorizer tests still pass.

Closes #2518.
